updated the backend for delete_file 4

+++ backend/src/controller/file.controller.js
@@ -114,6 +114,7 @@ const getAllFiles = (req, res) => {
 };
 
 const delete_file = async (req, res) => {
+
     console.log(req.body)
     const data = req.body;
     if (!data) {
@@ -125,24 +126,42 @@ const delete_file = async (req, res) => {
         .replace(`${protocol}://${req.get('host')}${basePath}/uploaded_files/`, "")
     const folder_name = file_name.replace('.pdf', '');
     console.log("folder name:" + folder_name)
-    const filePath = paths.join(process.cwd(), 'uploaded_files', folder_name);
+    const filePath = paths.join(process.cwd(), 'uploaded_files/digital_sign/', folder_name + "/" + file_name);
     console.log("file path:" + filePath)
 
-
-    fs.unlink(filePath, (err) => {
-        if (err) {
-            return res.status(500).json({ error: 'Unable to delete file' });
+    fs.exists(filePath, (exists) => {
+        if (!exists) {
+            return res.status(404).json({ error: 'File not found' });
         }
-        const folderPath = paths.dirname(filePath);
-
-        console.log("to_remove:" + filePath)
-        fs.rm(folderPath, { recursive: true, force: true }, (err) => {
+        fs.unlink(filePath, (err) => {
             if (err) {
-                return res.status(500).json({ error: 'Unable to delete folder' });
+                return res.status(500).json({ error: 'Unable to delete file' });
             }
-            res.json({ message: 'File and folder deleted successfully' });
+            res.json({ message: 'File deleted successfully' });
         });
     });
+    // const data = req.body;
+    // if (!data) {
+    //     return res.status(400).json({ error: 'File name is required' });
+    // }
+
+    // const protocol = getProtocol(req);
+    // const file_name = data.url
+    //     .replace(`${protocol}://${req.get('host')}${basePath}/uploaded_files/`, "")   
+
+    // const filePath = paths.join(process.cwd(), 'uploaded_files', file_name);
+
+    // fs.exists(filePath, (exists) => {
+    //     if (!exists) {
+    //         return res.status(404).json({ error: 'File not found' });
+    //     }
+    //     fs.unlink(filePath, (err) => {
+    //         if (err) {
+    //             return res.status(500).json({ error: 'Unable to delete file' });
+    //         }
+    //         res.json({ message: 'File deleted successfully' });
+    //     });
+    // });
 };
 
 const deletea = async (req, res) => {
